+++ Core/Domain/Clicco.Domain.Model/Transaction.cs
@@ -5,8 +5,8 @@ namespace Clicco.Domain.Model
     public class Transaction : BaseEntity, ISoftDeletable
     {
         public string Code { get; set; }
-        public int TotalAmount { get; set; }
-        public int DiscountedAmount { get; set; }
+        public  decimal TotalAmount { get; set; }
+        public  decimal DiscountedAmount { get; set; }
         public string Dealer { get; set; }
         public DateTime DeliveryDate { get; set; }
         public DateTime CreatedDate { get; set; }
